Fix Good_ bug of moving the wrong alias to tried
Advancing here doesnt work: we just set nLastSuccess and nLastTry
and would lose this information if we moved on to an alias and promoted
that to tried - the promoted entry would be Terrible from the
start.

Removing this line means that we'll promote the source of the canonical entry to tried
(and therefore would choose another bucket if we'd ever yeet back to
new). That's ok imo, and mirrors the existing behavior.