refactor: Split up FindBlockPos function
FindBlockPos() does different things depending on whether the block is known or not,
as shown by the fact that much of its old code was conditional on the fKnown flag.

If the block position is known (during reindex) the function would update the block info
statistics, so that we know where to put the next new block. It doesn't actually find a
block position in this case.

This commit splits up these two code paths by introducing a separate function
UpdateBlockFileInfo for the case when the block position is known.
It doesn't change behavior.