tests: add addrman_select_by_network test
this adds coverage for the 7 different cases of which table should be selected
when the network is specified. the different cases are the result of newOnly
being true or false and whether there are network addresses on both, neither,
or one of new vs tried tables. the only case not covered is when newOnly is
false and the only network addresses are on the new table.

Co-authored-by: Martin Zumsande <mzumsande@gmail.com>