net: increase inbound capacity for block-relay-only connections
..and adjust the eviction logic.
The new default max connection number is 200, the number of tx-relaying
inbounds is limited to the previous maximum inbound connections (114).
As a result, the tx-related maximum traffic does not change.

When we receive an inbound connection and don't have space for another
full-relay peer, we now evict only a full-relay inbound for it.

Once this commit widely deployed, the added inbound capacity will
allow us to increase the number of block-relay-only connections.

Co-authored-by: Amiti Uttarwar <amiti@uttarwar.org>