validation: add more checks to CheckBlockIndex
This adds checks that
1) descendants of failed blocks are also marked as failed
2) m_best_header points to the most-work, non-invalid block in our block
   index.

The added checks should be satisfied in normal node mode after #30666.
However, they can be broken in the case of a invalidateblockindex
rpc call, which will release cs_main before calling InvalidChainFound().

TBD: How to deal with that - we probably want InvalidateBlock() to stay
performant, and we do want to release cs_main - maybe we can apply some more cashing?

For now, use this commit only internally for fuzzing.